practice dir, add dayone.js and transfer first day functional component
